# RD-1295 Add TS project references