--- a/.eslintignore
+++ b/.eslintignore
@@ -6,5 +6,6 @@ doc/**
 templates/**
 tours/**
 userData/**
+**/dist/**
 
 # To be removed, when code is fixed in these folders