--- a/.eslintignore
+++ b/.eslintignore
@@ -6,6 +6,6 @@ doc/**
 templates/**
 tours/**
 userData/**
-**/dist/**
+tsc-dist/**
 
 # To be removed, when code is fixed in these folders