--- a/app/tsconfig.json
+++ b/app/tsconfig.json
@@ -4,7 +4,7 @@
         "types": [],
         "plugins": [{ "name": "typescript-styled-plugin" }],
         "composite": true,
-        "outDir": "dist",
+        "outDir": "../tsc-dist",
         "rootDir": ".."
     },
     "include": ["**/*", "**/*.json"],