--- a/backend/tsconfig.json
+++ b/backend/tsconfig.json
@@ -1,10 +1,5 @@
 {
-    "extends": "../node_modules/cloudify-ui-common/configs/tsconfig.composite.json",
+    "extends": "../tsconfig.base.json",
     "exclude": ["node_modules", "test", "dist"],
-    "include": ["**/*", "../conf/**/*.json"],
-    "compilerOptions": {
-        "rootDir": "..",
-        "composite": true,
-        "outDir": "../tsc-dist"
-    }
+    "include": ["**/*", "../conf/**/*.json"]
 }