--- a/package.json
+++ b/package.json
@@ -25,7 +25,7 @@
     "check-types": "npm run check-types:backend && npm run check-types:frontend && npm run check-types:widgets && npm run check-types:cypress",
     "//": "The backend and app directories are type-checked when compiling their respective test projects, so there's no need to type-check them individually",
     "check-types:backend": "tsc --build backend backend/test",
-    "check-types:frontend": "tsc -p test/jest --checkJs false",
+    "check-types:frontend": "tsc --build app test/jest",
     "check-types:widgets": "tsc -p widgets --checkJs false",
     "check-types:cypress": "tsc -p test/cypress --checkJs false",
     "coverageCheck": "cp coverage-jest/coverage-final.json .nyc_output && nyc report --reporter lcov --report-dir coverage-overall && nyc check-coverage",