--- a/package.json
+++ b/package.json
@@ -26,7 +26,7 @@
     "//": "The backend and app directories are type-checked when compiling their respective test projects, so there's no need to type-check them individually",
     "check-types:backend": "tsc --build backend backend/test",
     "check-types:frontend": "tsc --build app test/jest",
-    "check-types:widgets": "tsc -p widgets --checkJs false",
+    "check-types:widgets": "tsc --build widgets",
     "check-types:cypress": "tsc -p test/cypress --checkJs false",
     "coverageCheck": "cp coverage-jest/coverage-final.json .nyc_output && nyc report --reporter lcov --report-dir coverage-overall && nyc check-coverage",
     "devServer": "cross-env NODE_ENV=development node devServer.js",