--- a/package.json
+++ b/package.json
@@ -35,7 +35,7 @@
     "e2e": "cypress run -C node_modules/cloudify-ui-common/cypress/cypress.json",
     "e2e:open": "cypress open -C node_modules/cloudify-ui-common/cypress/cypress.json",
     "e2e:ci": "npm run beforebuild && cross-env TEST=1 run-system-tests ${TEST_SPEC}",
-    "lint": "eslint --cache --ignore-path .gitignore --ext js,jsx,ts,tsx --max-warnings 10 .",
+    "lint": "eslint --cache --ignore-path .gitignore --ext js,jsx,ts,tsx .",
     "size": "[ -d \"dist\" ] && size-limit",
     "test": "npm run test:frontend && npm run test:backend",
     "test:frontend": "jest test/jest",