--- a/test/cypress/integration/getting_started_spec.ts
+++ b/test/cypress/integration/getting_started_spec.ts
@@ -79,6 +79,7 @@ function interceptSecretsCreation(secrets: string[]) {
 
 function interceptPluginsUpload(plugins: string[]) {
     plugins.forEach(plugin => {
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         const catalogEntry = find(pluginsCatalog, { name: plugin })!;
         cy.intercept({
             method: 'POST',