--- a/test/cypress/integration/widgets/deployments_view_spec.ts
+++ b/test/cypress/integration/widgets/deployments_view_spec.ts
@@ -5,7 +5,7 @@ import type { SystemLabel } from '../../support/deployments';
 
 import { exampleBlueprintUrl } from '../../support/resource_urls';
 import { FilterRuleOperators, FilterRuleType } from '../../../../widgets/common/src/filters/types';
-import '../../../../widgets/common/src/deploymentsView';
+import type {} from '../../../../widgets/common/src/deploymentsView';
 
 describe('Deployments View widget', () => {
     const widgetId = 'deploymentsView';