--- a/tsconfig.base.json
+++ b/tsconfig.base.json
@@ -1,0 +1,9 @@
+{
+    "extends": "./node_modules/cloudify-ui-common/configs/tsconfig.composite.json",
+    "compilerOptions": {
+        "declarationMap": false,
+        "composite": true,
+        "outDir": "./tsc-dist",
+        "rootDir": "."
+    }
+}