--- a/tsconfig.json
+++ b/tsconfig.json
@@ -1,11 +1,10 @@
 {
     "references": [
-        { "path": "app" },
+        { "path": "./tsconfig.ui.json" },
         { "path": "backend" },
         { "path": "backend/test" },
         { "path": "test/jest" },
-        { "path": "test/cypress" },
-        { "path": "widgets" }
+        { "path": "test/cypress" }
     ],
     "files": []
 }