--- a/tsconfig.ui.json
+++ b/tsconfig.ui.json
@@ -1,0 +1,4 @@
+{
+    "references": [{ "path": "app" }, { "path": "widgets" }],
+    "files": []
+}