--- a/webpack.config.js
+++ b/webpack.config.js
@@ -217,7 +217,7 @@ module.exports = (env, argv) => {
                             }
                         },
                         typescript: {
-                            configFile: './tsconfig.json',
+                            configFile: './tsconfig.ui.json',
                             build: true,
                             mode: 'write-references'
                         }