--- a/widgets/common/src/filters/AddRuleButton.tsx
+++ b/widgets/common/src/filters/AddRuleButton.tsx
@@ -1,13 +1,12 @@
 import type { ComponentProps, FunctionComponent } from 'react';
 import { getTranslation } from './common';
 
-const { Button } = Stage.Basic;
-
 interface AddRuleButtonProps {
     onClick: ComponentProps<typeof Stage.Basic.Button>['onClick'];
 }
 
 const AddRuleButton: FunctionComponent<AddRuleButtonProps> = ({ onClick }) => {
+    const { Button } = Stage.Basic;
     return <Button icon="add" content={getTranslation('buttons.addRule')} onClick={onClick} />;
 };
 export default AddRuleButton;