--- a/widgets/common/src/labels/AddButton.tsx
+++ b/widgets/common/src/labels/AddButton.tsx
@@ -1,7 +1,5 @@
 import type { ComponentProps, FunctionComponent } from 'react';
 
-const { Button } = Stage.Basic;
-
 interface LabelAddButtonProps {
     disabled: ComponentProps<typeof Stage.Basic.Button>['disabled'];
     onClick: ComponentProps<typeof Stage.Basic.Button>['onClick'];
@@ -9,7 +7,10 @@ interface LabelAddButtonProps {
 }
 
 const LabelAddButton: FunctionComponent<LabelAddButtonProps> = ({ disabled, onClick, onEnterPress }) => {
-    const { i18n } = Stage;
+    const {
+        i18n,
+        Basic: { Button }
+    } = Stage;
 
     function handleKeyDown({ key }: { key: string }) {
         if (key === 'Enter') onEnterPress();