--- a/widgets/common/src/labels/LabelErrorPopup.tsx
+++ b/widgets/common/src/labels/LabelErrorPopup.tsx
@@ -1,12 +1,11 @@
 import type { ComponentProps, FunctionComponent } from 'react';
 
-const { Popup } = Stage.Basic;
-
 interface LabelsErrorPopupProps {
     content: ComponentProps<typeof Stage.Basic.Popup>['content'];
 }
 
 const LabelsErrorPopup: FunctionComponent<LabelsErrorPopupProps> = ({ content }) => {
+    const { Popup } = Stage.Basic;
     return <Popup open trigger={<div />} content={content} position="top left" pinned wide />;
 };
 export default LabelsErrorPopup;