… widgets/tsconfig.json