This PR adds support for [project references](https://www.typescriptlang.org/docs/handbook/project-references.html) in our TS projects.

The projects structure is as follows:

![image](https://user-images.githubusercontent.com/889383/120282991-fc728000-c2ba-11eb-9ae7-12bbcf16e40f.png)

and then there is the main `tsconfig.json` solution project that includes all of the projects. Arrow `A -> B` means that project `A` uses `B` as a reference.

TypeScript compilation results (only declaration files and declaration maps, as well as a `tsbuildinfo` file) are cached in the `tsc-dist` directory. They are not used in any user-facing solution.

## Benefits

1. Errors from unrelated projects can be shown at the same time in webpack/`npm run check-types`

    For example, if there were errors in both `test/jest` and `test/cypress` and `backend`, all those projects do not depend on each other, so they can be type-checked in parallel. Previously, due to using `&&` in the `npm run check-types` command, only errors from a single project were shown.

2. TypeScript detects references across projects.

    For example, when searching for references to some `Stage` API declared in `app`, IDEs should now correctly detect the use of that API in other projects, e.g. `widgets`.

    With project references: 
    ![image](https://user-images.githubusercontent.com/889383/120285201-507e6400-c2bd-11eb-9be9-08cccc2df9a5.png)

    Without project references: 
    ![image](https://user-images.githubusercontent.com/889383/120286153-3d1fc880-c2be-11eb-8954-211569701c42.png)



3. Webpack reports up-to-date errors when API in `app` changes.

    If there is an API change in `app`, webpack recompiles the types in `app` and then recompiles `widgets`. Previously, webpack relied on the initial version of `app`, meaning that if you changed some `app` API, errors in `widgets` due do that API change were not reported.

    For example, I replaced `StageUtils.isEmptyWidgetData` with `null`.
    With project references: 
    ![image](https://user-images.githubusercontent.com/889383/120285620-b79c1880-c2bd-11eb-86db-2c9fb396f7f9.png)

    Without project references, webpack is happy:
    ![image](https://user-images.githubusercontent.com/889383/120286025-22e5ea80-c2be-11eb-8178-2c146dbbbea7.png)


4. Less resource usage in IDEs

    IDEs can now use a single TS language server for the main `tsconfig.json` solution file instead of starting up a new TS language server for each subproject. This means that when working in multiple directories (and thus, multiple projects), you will have just a single TS language server and will compile each file only once, instead of N language servers and N compilations.

    With project references (a single TS language server): 
    ![image](https://user-images.githubusercontent.com/889383/120285266-5d02bc80-c2bd-11eb-814d-f223037d3d5f.png)

    Without project references (multiple TS servers):
    ![image](https://user-images.githubusercontent.com/889383/120285922-0944a300-c2be-11eb-8f7f-fedcf553983f.png)

5. Better performance when type-checking (see [Performance](#performance))

## Code changes

Any code changes done in this PR are due to the `declarationMap` setting being enabled in TS projects. Because of it there can be no private types used in public APIs. I had to modify some files that had unusual code structure, e.g. did not export the main component, or used some private variable to extend some global API.

Feel free to comment on specific changes and I will try to provide explanations, but perhaps the fastest way would be to try to revert the change locally and inspect the error from `npm run check-types` :slightly_smiling_face: 

## Not addressed in this PR

I have not addressed the following items in this PR:
1. Configure path aliases to shorten import paths. This would introduce even more changes that are not strictly related to project references. Project references support turned out to be hard enough IMO that configuring import paths can be done in a separate ticket.
    I have modified RD-1724 (an existing ticket for import paths) to capture this requirement
2. Upgrade to TypeScript 4.3. It introduces some improvements for TS project references performance and files size. I will do that in a separate PR, still as a part of RD-1295

## Performance

Performance comparison of type-checking is available here:
https://docs.google.com/spreadsheets/d/1rMQYPDzRhN44GER2TXPcvnBBszUUodZNK-w11IS4jAE/edit#gid=0

Conclusion:
1. Rechecking in webpack may be now slower, since webpack potentially has to recompile 2 projects, not just 1 (see benefit number 3).
2. `npm run check-types` is 50% slower when doing a cold start (without the `tsc-dist` cache) (30s vs 19s).
3. `npm run check-types` is 98% or 50% faster with a warm cache (0.45s when no changes, 8.18s when there are changes, compared to 19s without project references)

I hope all of the above are improved with the TS 4.3 upgrade based on these premises:
1. https://devblogs.microsoft.com/typescript/announcing-typescript-4-3/#tsbuildinfo-is-smol
2. https://devblogs.microsoft.com/typescript/announcing-typescript-4-3/#lazier-incremental

## System tests

https://jenkins.cloudify.co/blue/organizations/jenkins/Stage-UI-System-Test/detail/Stage-UI-System-Test/714/pipeline

Queued 2021-06-07 9:00 CEST