@Gelio 2021-06-01
I believe I have fixed the system tests - we needed to use `import type` in 2 places that I added in this PR.

Here's a rerun:
https://jenkins.cloudify.co/blue/organizations/jenkins/Stage-UI-System-Test/detail/Stage-UI-System-Test/695/pipeline