@Gelio 2021-06-01
@qooban Sure, no worries, I believe a single approval could be enough, as most of the changes are related to the configuration and then addressing any TS errors the came up

@kubama @qooban Can either of you check if the project still works well in WebStorm, if you can go to references, etc? I don't want to break the development experience for you :smile: