@Gelio 2021-06-02
This PR had conflicts anticipated in https://github.com/cloudify-cosmo/cloudify-stage/pull/1404#discussion_r642871222 since #1402 was merged.

You can see the merge conflicts resolution by running:

```sh
git show e9f1d1bd2382d7faae0f859db4a077227b36499a
```


`npm run check-types` still works after resolving those. However, I'll run a full system test suite just in case :smile: 

https://jenkins.cloudify.co/blue/organizations/jenkins/Stage-UI-System-Test/detail/Stage-UI-System-Test/706/pipeline