@Gelio 2021-06-07
Looks like this can finally be merged :smile: 

I just got 2 failures in the e2e tests, both in the Getting Started modal, both coming from some incompatibility with some AWS blueprint:

![image](https://user-images.githubusercontent.com/889383/120991556-0a744500-c782-11eb-90dd-907f0752effc.png)

:worried:

I don't think it's a blocker for this PR, since it does not seem to be related, but anyway, that should probably be fixed soon. I'll create a bug for that and post it on Slack, and I'll merge the PR in the meantime